Obtener la configuración del usuario por su Id

+++ api/src/main/java/com/hosting/rest/api/services/User/UserConfiguration/UserConfigurationServiceImpl.java
@@ -1,12 +1,18 @@
 package com.hosting.rest.api.services.User.UserConfiguration;
 
+import static com.hosting.rest.api.Utils.AppUtils.isNotNull;
+import static com.hosting.rest.api.Utils.AppUtils.isIntegerValidAndPositive;
+
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
+import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.hosting.rest.api.exceptions.IllegalArguments.IllegalArgumentsCustomException;
+import com.hosting.rest.api.exceptions.NotFound.NotFoundCustomException;
 import com.hosting.rest.api.models.User.UserConfiguration.UserConfigurationModel;
 import com.hosting.rest.api.repositories.User.UserConfiguration.IUserConfigurationRepository;
 
@@ -21,27 +27,56 @@ public class UserConfigurationServiceImpl implements IUserConfigurationService {
 
 	@Override
 	public UserConfigurationModel addNewUserConfiguration(final UserConfigurationModel newUserConfigurationModel) {
-		// TODO: Comprobar parametro
+		if (!isNotNull(newUserConfigurationModel)) {
+			throw new IllegalArgumentsCustomException(
+					"Alguno de los valores de la configuración a añadir no es válido.");
+		}
+
 		return userConfigurationRepo.save(newUserConfigurationModel);
 	}
 
+	@Transactional
 	@Override
 	public UserConfigurationModel updateUserConfiguration(final Integer userId,
-			UserConfigurationModel userConfigurationModelToUpdate) {
-		// TODO: Comprobar parametro
+			final UserConfigurationModel userConfigurationModelToUpdate) {
 
+		if (!isIntegerValidAndPositive(userId)) {
+			throw new IllegalArgumentsCustomException("El id del usuario [ " + userId + " ] no es válido.");
+		}
+
+		if (!isNotNull(userConfigurationModelToUpdate)) {
+			throw new IllegalArgumentsCustomException(
+					"Alguno de los valores de la configuración introducidos no es válido.");
+		}
+
+		// Comprobar si existe una configuación para el usuario
+		boolean existsUserConfiguration = userConfigurationRepo
+				.existsById(userConfigurationModelToUpdate.getIdUserConfiguration());
+
+		if (!existsUserConfiguration) {
+			throw new NotFoundCustomException("La configuración del usuario a actualizar no existe.");
+		}
+
+		// TODO: Actualizar datos de la configuración
+		
 		return null;
 	}
 
 	@Override
 	public UserConfigurationModel findById(final Integer userConfigurationId) {
-		// TODO: Comprobar parametro
+		if (!isIntegerValidAndPositive(userConfigurationId)) {
+			throw new IllegalArgumentsCustomException(
+					"El id de la configuración del usuario [ " + userConfigurationId + " ] no es válido.");
+		}
+
 		return userConfigurationRepo.findById(userConfigurationId).get();
 	}
 
 	@Override
 	public UserConfigurationModel findByUserId(final Integer userId) {
-		// TODO: Comprobar parametro
+		if (!isIntegerValidAndPositive(userId)) {
+			throw new IllegalArgumentsCustomException("El id del usuario [ " + userId + " ] no es válido.");
+		}
 
 		String findByUserIdQuery = "SELECT ucm " + "FROM UserModel um INNER JOIN um.idUserConfiguration ucm "
 				+ "WHERE um.id = :userId";
@@ -54,21 +89,14 @@ public class UserConfigurationServiceImpl implements IUserConfigurationService {
 	}
 
 	@Override
-	public void deleteUserConfigurationByUserId(final Integer userId) {
-		// TODO: Comprobar parametro
-		String findByUserIdQuery = "DELETE FROM UserConfigurationModel ucm, UserModel um "
-				+ "WHERE um.id = :userId AND um.idUserConfiguration.id = ucm.id";
+	public void deleteUserConfiguration(final Integer userConfigurationId) {
+		if (!isIntegerValidAndPositive(userConfigurationId)) {
+			throw new IllegalArgumentsCustomException(
+					"El id de la configuración del usuario [ " + userConfigurationId + " ] no es válido.");
+		}
 
-		TypedQuery<UserConfigurationModel> userConfig = em.createQuery(findByUserIdQuery, UserConfigurationModel.class);
-
-		userConfig.setParameter("userId", userId);
-	}
-
-	@Override
-	public void deleteUserConfiguration(Integer userConfigurationId) {
-		// TODO: Comprobar parámetro
 		userConfigurationRepo.deleteById(userConfigurationId);
-		
+
 	}
 
 }
